Refactor Field.hpp add .cpp file for include controle
* Added narrowing functions to Util with proper checks (overflow).
* Add possibility to silence error logging e.g. If you expect an uint8_t value from database but the value is negative, it will report the issue now. If you are aware of that and want to narrow it without error to numeric_limit, you can silence it.
* cleanup includes in project.
* Todo:
-- struct ItemProperties, uint32_t ArmorDamageModifier, there are float values in database (negative as well).  Changed to float
-- struct CreatureSpawn changed channel_spell from uint16_t to uint32_t since WoWUnit sends uint32_t anyway.
-- corrected data loading for reputation_faction_onkill in function loadReputationModifierTable. It is not the same format

+++ src/world/Management/ItemProperties.hpp
@@ -80,7 +80,7 @@ struct ItemProperties
     uint32_t SocketBonus;
     uint32_t GemProperties;
     int32_t DisenchantReqSkill;
-    uint32_t ArmorDamageModifier;
+    float ArmorDamageModifier;
     uint32_t ExistingDuration;
     uint32_t ItemLimitCategory;
     uint32_t HolidayId;
